requirements.txt
pytest
pytest-playwright
playwright
pytest-html
pyyaml
requests
faker